Add files via upload
Code of Draw2Measure Protractor, an iOS app developed by American Printing House for the Blind that gives students an alternative way to measure angles.

The app allows students to measure angles in two ways:

First, students can place an angle over the screen of a device, such as a phone or tablet, and trace along the sides of the angle with a fingertip or stylus. The app records the locations of the sides and then calculates the angle. 

For objects that may not fit on a screen, students can find measurements by rotating the device itself, which utilizes a built-in gyroscope sensor to measure the angle.